move refinement tests into query tests
